Added nasty workaround for Fedora JeOS
In config.sh two lines to copy the kernel image(vmlinuz)
and the System.map to /boot has been added. Actually this
was not required before and I'm sure I don't understand
the concept behind the kernel packaging at Fedora, which
makes this a workaround I'd like to get rid of as soon
as possible.

When dnf install the kernel package (kernel-core) the files
/boot/System.map and also /boot/vmlinuz are listed as to
belong to the package. But those files does not exist on
the system nor does rpm -V complain about their absence.
Thus I guess they are marked as %ghost in the spec file.

I thought the kernel-install post script calls would care
for the kernel files to become copied from /lib/modules/...
to /boot but that did not happen either, which leads me
adding the two copy lines in config.sh of the image

Again this is not what I want to do but it was the only
way to make the image build and boot again. I'm open for
any better fix