Fixup console multiplexer
In case multiple consoles are configured plymouth usually
multiplexes the output to all open consoles. This is true
until kiwi calls the dialog program to run interactive
dialogs. Interaction requires keyboard input which is
occupied by plymouth on the master console. Therefore the
custom kiwi boot code calls the hide-splash command which
tells plymouth to release this resource. As consequence
you also loose the console multiplexer. Therefore this
patch only calls hide-splash if there is only one active
tty such that we can be sure the user has a chance to
interact with dialogs.