Use setctsid/bash instead of sulogin for debugging
The custom kiwi boot descriptions supports a boot process
accompanying debug shell if the kiwidebug=1 option is
set on the kernel command line. That shell was invoked
using sulogin. Problem is that this call occupies the
controlling terminal and console such that the typical
multiplexing of consoles through plymouth is stopped.
This means it's not possible to run e.g an oem deployment
with several active/multiplexed consoles and a backgrounding
debuging shell. This patch moves away from sulogin and
calls a bash through setctsid on a free console