Use default values for nr_inode on devtmpfs mounts
Using nr_inode=0 is causing mount failures in recent kernels. This
option was introduced to fix issues in split systems, which are no
longer supported. According to that we just remove this option and
let mount set the default value, which depends on available memory
of the system.

Fixes SUSE/kiwi#1105