Added kiwi-image:wsl Provides
For support of the wsl/appx container format a new image
type will be provided. The tool to create that image type
is packaged in fb-util-for-appx. To allow the buildservice
to build this image type the obs worker has to install
this tool. This commit creates a sub-package for use with
the buildservice that manages all needed packages within
the new kiwi-image:wsl provides tag. This is related to
https://github.com/OSInside/kiwi/issues/1235