Delete custom initrds for any arch except x86_64
With the move to dracut kiwi's custom initrd support became
obsolete. After announcing this with SLE12-SP4 the transition
period is now over as SLE12 reached EOL. With the exception
of the x86_64 architecture there is no other architecture
using custom kiwi initrds (netboot for SLE12 on s390 was a
technology preview which is no longer in use). Thus the
support for other than the x86_64 architecture will be phased
out with this commit.