Update descriptions for use with one disk type
As we have changed kiwi to support the vmx type as oem type
with rootfs resize disabled, because this is the only difference
between the two, the image descriptions that defines oem and
vmx types in one preferences section needs to be adapted.
The automatic transformation from the vmx type into an oem
type can lead to two oem definitions in one preferences
section and at build time the last one will win. This commit
updates the image descriptions that uses such type
definitions and organize them in profiles.