Keep only one version of netboot for reference
netboot is a legacy custom kiwi boot description. With the
drop of the kiwi-boot-descriptions package in request 863256
we finally stop maintaining this code in the scope of the
appliance building team. People who still use netboot have
created their own copy of it and maintain it in their space.
For reference we keep one code base here but I expect this
one to go as well after the dust has settled. This references
the history of Issue OSInside/kiwi#1299