Added suse-leap-15.3-extra-lvm
A system with LVM as add-on not containing the root
filesystem. The LVM is placed in an extra partition at
the end of the disk such that it could also be resized.
The description uses the spare_part feature and since
the extra LVM partition is last and not root you are
on your own regarding the resize. However I can imagine
existing grow-part solutions (e.g cloud-init) can do
the trick